Sending bids and managing response statuses

+++ src/main/java/com/cars24/solution/db/BiddingRepository.java
@@ -9,5 +9,5 @@ public interface BiddingRepository extends PagingAndSortingRepository<BiddingSys
 
  // List<BiddingSystem> findAll();
 
-  BiddingSystem findById(long id);
+  BiddingSystem findById(int id);
 }
